Support CSS motion path animations
CSS motion path animations allow an element to be animated along a
path. This is usually done by setting the offset-path property, and
then animating offset-distance property. For the current time being,
only the path function is supported for offset-path property.
Additionally, the offset-rotate property can be used to control the
angle along the path. There are five different modes which this change
support.

There are some unsupported properties that can be added later including
the short hand form for defining a motion path animation using the
offset property.

Fixes: QTBUG-137859
Change-Id: I51559f378dc4397786d71de1ce006ae1df4352e9
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>