Don't render QSvgPattern for calculating viewBox
Instead of the pattern rendered into a QImage just return a default
QImage of the same size. For calculating the bounds, the image's
contents don't matter.

While this speeds up loading of images which don't set a viewBox
themselves, the main purpose of this change is to avoid endless
recursions. The image's viewBox is being calculated before QSvgHandler
tries to detect cycles, so the program could run into a stack overflow.

This accompanies 935dcfcf4118fe8eb2fb1c64a06923939379dfce which caught
these recursions later. Nevertheless that change cannot be reverted now
because it catches further cases.

Change-Id: I2039fb3d6155b154713d5bcab380025f95665c68
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>