Add an example to the module
The module used to have an example to view SVG files. However,
it was using the Graphics View framework and thus moved to manual
tests. This new example is a reworked version of the old one, and it
is based entirely on Qt Widgets. Moreover, new features were added to
make it more appealing and to expose the new capabilities introduced
by the module.

New Features :
- Features menu to enable and disable the newly added features to
  the module
- Three background options
- Refresh button and Auto Refresh option to automatically re-render
  a file that has been modified.

Task-number: QTBUG-135261
Pick-to: 6.10
Change-Id: I6cc3b0474d770990630a27e2808f969d16a26118
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>