Fix parsing of url defined by CSS
Changes done in 6065c5fede827a8bd595f18c017b2a549fbb28c0 fixed an
issue with parsing multiple values in a CSS style sheet,
but introduced a bug where the url is not read at all. This caused
a regression when a url to a gradient for example is used in CSS.

Parsed CSS values in Qt Svg were never checked for validity, so this
commit also introduces a new test file that can be extended later
to test the parsed values by QSvgCssHandler.

Fixes: QTBUG-140951
Pick-to: 6.10
Change-Id: Ia032fac709b1e03fef47810f886e4f813ab07144
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>