Add support for CSS timing function basic values
An animation can specify a timing function that can be used for
easing. CSS supports a wide range of timing functions, but this
commit focuses on creating the base implementation for supporting
easing in the module. This commit implements the "keywords"
timing function. This includes linear, ease, ease-in, ease-out,
ease-in-out, step-start, and step-end.

Task-number: QTBUG-139619
Change-Id: Ic5ca2e5249a95811ef9846c3ecaf0c3fdaf16604
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>