Avoid exponential growth over benchmark's iterations
One of the benchmarks draws into a QPicture. A QPicture adjusts its
bounding rectangle when drawing to it unless the rectangle was set
explicitly. On the next iteration that rectangle is then used in
QSvgTinyDocument::mapSourceToTarget to calculate the renderings scaling
factor. That feedback loop can lead to an exponential growth of memory
and time consumption as it happened here with filter.svg.

Setting a bounding rectangle beforehand breaks the feedback loop and
keeps the benchmarks' run times linear to the number of iterations.

Task-number: QTBUG-141915
Change-Id: I2a07a96aace828001d24db7468484a4b3e41b6e6
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
(cherry picked from commit 99babcbafcffd9eb6184f41164ae4a2baa50ef48)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>