QSvgFeGaussianBlur: Read and write image data only once
The old code calculated each color component (r,g,b,a) in a separate
pass which meant that the entire processed image was read and written
four times in a row. The new code now does the calculations for all
color components of a pixel right after one another so the raw image
data needs to be read and written only once which allows better caching
of the data.

Pick-to: 6.11
Change-Id: Iaee021fa109438cb56eab84b0a9dfbb4fb5f1152
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>