Fix default color handling in parseStopNode()
The handling of default color changed by the commit mentioned below
introduced a regression: Whenever the found color string didn't define
an actual color, black was used as the default, but a possibly valid
opacity was ignored. This patch adds code to assign the requested
opacity to the default color as it was already done in other places.

The patch contains a baseline test which would have shown the
regression.

Cherry-pick goes to 6.11 which is the oldest branch which contains the
breaking change.

Amends 44f83cdadde01b638235660272d3b3602f7931c5

Fixes: QTBUG-142727
Pick-to: 6.11
Change-Id: I97f648fe08c8d10fa2f27b4beb87657061da8646
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>